Extract virtualenv test case form venv test
